refactor(compression): order codec subpaths by direction
