+++ modules/compression/README.md
@@ -6,7 +6,8 @@ compact JavaScript, and optional codec implementations.
 ## Design
 
 - Async codecs prefer the runtime's `CompressionStream` and `DecompressionStream` support.
-- Direction-specific subpaths let decode-only applications exclude encoders, and vice versa.
+- Direction-specific subpaths use `FORMAT-DIRECTION-IMPLEMENTATION` order and let decode-only
+  applications exclude encoders, and vice versa.
 - Compact `fflate`, `fzstd`, Snappy, LZ4 block, and Brotli fallbacks cover common read paths.
 - Larger or specialized implementations are selected through explicit, independently importable
   subpaths.
@@ -16,7 +17,7 @@ compact JavaScript, and optional codec implementations.
 ## Example
 
 ```typescript
-import {GZipFflateDecompressor} from '@loaders.gl/compression/gzip-fflate-decompressor';
+import {GZipFflateDecompressor} from '@loaders.gl/compression/gzip-decompressor-fflate';
 
 const decompressor = new GZipFflateDecompressor();
 const data = await decompressor.decompress(compressedData);
