+++ modules/compression/src/bzip2-compressor-compress-utils.ts
