+++ modules/compression/src/bzip2-decompressor-compress-utils.ts
