+++ modules/compression/src/deflate-compressor-compress-utils.ts
