+++ modules/compression/src/deflate-compressor-pako.ts
