+++ modules/compression/src/gzip-compressor-compress-utils.ts
