+++ modules/compression/src/gzip-compressor-fflate.ts
