+++ modules/compression/src/gzip-compressor-pako.ts
