+++ modules/compression/src/lz4-compressor-compress-utils.ts
