+++ modules/compression/src/snappy-compressor-compress-utils.ts
