+++ modules/compression/src/snappy-decompressor-compress-utils.ts
