+++ modules/compression/test/library-bindings.spec.ts
@@ -8,23 +8,23 @@ import {NoCompressor} from '@loaders.gl/compression/no-compressor';
 import {NoDecompressor} from '@loaders.gl/compression/no-decompressor';
 import {DeflateFflateCompression} from '@loaders.gl/compression/deflate-fflate';
 import {GZipFflateCompression} from '@loaders.gl/compression/gzip-fflate';
-import {GZipFflateCompressor} from '@loaders.gl/compression/gzip-fflate-compressor';
-import {GZipFflateDecompressor} from '@loaders.gl/compression/gzip-fflate-decompressor';
+import {GZipFflateCompressor} from '@loaders.gl/compression/gzip-compressor-fflate';
+import {GZipFflateDecompressor} from '@loaders.gl/compression/gzip-decompressor-fflate';
 import {DeflatePakoCompression} from '@loaders.gl/compression/deflate-pako';
 import {GZipPakoCompression} from '@loaders.gl/compression/gzip-pako';
-import {GZipPakoCompressor} from '@loaders.gl/compression/gzip-pako-compressor';
-import {GZipPakoDecompressor} from '@loaders.gl/compression/gzip-pako-decompressor';
+import {GZipPakoCompressor} from '@loaders.gl/compression/gzip-compressor-pako';
+import {GZipPakoDecompressor} from '@loaders.gl/compression/gzip-decompressor-pako';
 import {LZ4JSCompression} from '@loaders.gl/compression/lz4-lz4js';
 import {SnappyJSCompression} from '@loaders.gl/compression/snappy-snappyjs';
-import {ZstdFzstdDecompressor} from '@loaders.gl/compression/zstd-fzstd';
+import {ZstdFzstdDecompressor} from '@loaders.gl/compression/zstd-decompressor-fzstd';
 import {ZstdCodecCompression} from '@loaders.gl/compression/zstd-zstd-codec';
-import {BrotliLoadersGLDecompressor} from '@loaders.gl/compression/brotli-loaders-gl-decompressor';
+import {BrotliShimDecompressor} from '@loaders.gl/compression/brotli-decompressor-shim';
 import {BrotliCompressUtilsCompression} from '@loaders.gl/compression/brotli-compress-utils';
 import {BZip2CompressUtilsCompression} from '@loaders.gl/compression/bzip2-compress-utils';
 import {DeflateCompressUtilsCompression} from '@loaders.gl/compression/deflate-compress-utils';
 import {GZipCompressUtilsCompression} from '@loaders.gl/compression/gzip-compress-utils';
-import {GZipCompressUtilsCompressor} from '@loaders.gl/compression/gzip-compress-utils-compressor';
-import {GZipCompressUtilsDecompressor} from '@loaders.gl/compression/gzip-compress-utils-decompressor';
+import {GZipCompressUtilsCompressor} from '@loaders.gl/compression/gzip-compressor-compress-utils';
+import {GZipCompressUtilsDecompressor} from '@loaders.gl/compression/gzip-decompressor-compress-utils';
 import {LZ4CompressUtilsCompression} from '@loaders.gl/compression/lz4-compress-utils';
 import {SnappyCompressUtilsCompression} from '@loaders.gl/compression/snappy-compress-utils';
 import {XZCompressUtilsCompression} from '@loaders.gl/compression/xz-compress-utils';
@@ -46,17 +46,17 @@ describe('compression library bindings', () => {
     await expectRoundTrip(compression);
   });
 
-  test('zstd-fzstd decodes zstd-codec output', async () => {
+  test('zstd-decompressor-fzstd decodes zstd-codec output', async () => {
     const compressed = await new ZstdCodecCompression().compress(copyArrayBuffer(TEST_BYTES));
     const output = await new ZstdFzstdDecompressor().decompress(compressed);
     expect(new Uint8Array(output)).toEqual(TEST_BYTES);
   });
 
-  test('brotli-loaders-gl decodes compress-utils output', async () => {
+  test('brotli-decompressor-shim decodes compress-utils output', async () => {
     const compressed = await new BrotliCompressUtilsCompression().compress(
       copyArrayBuffer(TEST_BYTES)
     );
-    const output = await new BrotliLoadersGLDecompressor().decompress(compressed);
+    const output = await new BrotliShimDecompressor().decompress(compressed);
     expect(new Uint8Array(output)).toEqual(TEST_BYTES);
   });
